Update project architecture frame and java view controller Update toString method (Question java class)
