+++ MyQCM/app/src/main/java/fr/iia/cdsmat/myqcm/view/menu/LegalMentionFragment.java
@@ -1,4 +1,4 @@
-package fr.iia.cdsmat.myqcm.view;
+package fr.iia.cdsmat.myqcm.view.menu;
 
 
 import android.os.Bundle;
@@ -12,10 +12,10 @@ import fr.iia.cdsmat.myqcm.R;
 /**
  * A simple {@link Fragment} subclass.
  */
-public class ProfilFragment extends Fragment {
+public class LegalMentionFragment extends Fragment {
 
 
-    public ProfilFragment() {
+    public LegalMentionFragment() {
         // Required empty public constructor
     }
 
@@ -24,7 +24,7 @@ public class ProfilFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_profil, container, false);
+        return inflater.inflate(R.layout.fragment_legal_mention, container, false);
     }
 
 }
