+++ MyQCM/app/src/main/res/layout/app_bar_menu.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="fr.iia.cdsmat.myqcm.view.MenuActivity">
+    tools:context=".view.menu.MenuActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
