+++ MyQCM/app/src/main/res/layout/fragment_help.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fr.iia.cdsmat.myqcm.view.HelpFragment">
+    tools:context=".view.menu.HelpFragment">
 
     <Button
         android:id="@+id/nav_btnHelp"
